requirements.txt
Flask
python-dotenv
groq
flask_sqlalchemy
werkzeug